rag/requirements.txt
requests
psycopg2-binary
python-dotenv
psycopg2-binary>=2.9
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=1.13.0
requests>=2.28
python-dotenv>=1.0.0
numpy>=1.23